cmd: Changed the main logic to not rely on terraform-website
It has been migrated to registry docs and those docs follow a subscribe format and then registry pulls from there.

So now we have to pull from registry to konw the providers and then manually pull all the providers we want,
so then we can run the code to abstract the JSON from the docs.

The format of the docs also changed (just the file location), now they live in 'docs/' and they have no
sidebar and the categories are in the resource config file metadata.

Also not all Providers follow the new format so now we have 3 formats:
* Old
* New
* Unk: Which is like an inbetween of Old->New